% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{dskeleton}
\alias{dskeleton}
\title{skeleton distance calculation
, , , , ,}
\usage{
dskeleton(nnc1, nnc2, px1, px2, skeleton)
}
\arguments{
\item{nnc1}{a vector of the indexes of the two nearest neighbors of one point.}

\item{nnc2}{a vector of the indexes of the two nearest neighbors of another point.}

\item{px1}{the projection proportion of point 1 onto the edge by its two nearest knots, with the lower index knot as the origin.}

\item{px2}{the projection proportion of point 2 onto the edge by its two nearest knots, with the lower index knot as the origin.}

\item{skeleton}{a list returned by the voronSkeleton method constructing the skeleton graph.}
}
\value{
skeleton-based distance between two points on the skeleton
}
\description{
skeleton distance calculation
, , , , ,
}
