% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Mickey_data}
\alias{Mickey_data}
\title{Generating Mickey data}
\usage{
Mickey_data(n1 = 1000, n2 = 100, s1 = 1, s2 = 0.2, d = 2, sd_high = 0.1)
}
\arguments{
\item{n1}{a number indicating the number of points in the larger circle}

\item{n2}{a number indicating the number of points in each of the two smaller circles}

\item{s1}{a number indicating the base radius for the larger circle}

\item{s2}{a number indicating the base radius for each of the two smaller circles}

\item{d}{a number indicating the dimension of the generated data. d>2 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data matrix
\item clus - The vector of true cluster labels for each data point
}
}
\description{
Generating Mickey data
}
\examples{
X = Mickey_data()$data
plot(X[,1], X[,2])
}
