% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MixMickey1000}
\alias{MixMickey1000}
\title{Simulated Mix Mickey Data with dimension 1000}
\format{
A list with 2 elements:
\describe{
\item{data}{A data frame with 3200 rows (observations) and 1000 columns (variables)}
\item{clus}{A vector of length 3200 specifying the true cluster memberships}
}
}
\usage{
MixMickey1000
}
\description{
Simulated Mix Mickey Data with dimension 1000
}
\keyword{datasets}
