% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{MixMickey_data}
\alias{MixMickey_data}
\title{Generating MixMickey data}
\usage{
MixMickey_data(n1 = 2000, n2 = 600, s1 = 2, s2 = 1, d = 2, sd_high = 0.1)
}
\arguments{
\item{n1}{a number indicating the number of points in the larger Gaussian cluster}

\item{n2}{a number indicating the number of points in each of the two smaller Gaussian clusters}

\item{s1}{a number indicating the base radius for the larger Gaussian cluster}

\item{s2}{a number indicating the base radius for each of the two smaller Gaussian clusters}

\item{d}{a number indicating the dimension of the generated data. d>2 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data matrix
\item clus - The vector of true cluster labels for each data point
}
}
\description{
Generating MixMickey data
}
\examples{
X = MixMickey_data()$data
plot(X[,1], X[,2])
}
