% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MixStar1000}
\alias{MixStar1000}
\title{Simulated Mix Star Data with dimension 1000}
\format{
A list with 2 elements:
\describe{
\item{data}{A data frame with 3000 rows (observations) and 1000 columns (variables)}
\item{clus}{A vector of length 3000 specifying the true cluster memberships}
}
}
\usage{
MixStar1000
}
\description{
Simulated Mix Star Data with dimension 1000
}
\keyword{datasets}
