% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{MixStar_data}
\alias{MixStar_data}
\title{Generating MixStar data}
\usage{
MixStar_data(n = 1000, d = 2, sd_high = 0.1)
}
\arguments{
\item{n}{a number indicating the number of points in each of the Gaussian clusters}

\item{d}{a number indicating the dimension of the generated data. d>2 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data matrix
\item clus - The vector of true cluster labels for each data point
}
}
\description{
Generating MixStar data
}
\examples{
X = MixStar_data()$data
plot(X[,1], X[,2])
}
