% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{MM_data}
\alias{MM_data}
\title{Generating Manifold Mixture data}
\usage{
MM_data(
  n1 = 800,
  n2 = 400,
  n3 = 2000,
  sd1 = 0.04,
  sd2 = 0.04,
  sd3 = 0.04,
  d = 3,
  sd_high = 0.1
)
}
\arguments{
\item{n1}{number of points in the ring structure}

\item{n2}{number of points in the perturbed line segment structure}

\item{n3}{number of points in the 2-dimensional plane structure}

\item{sd1}{the standard deviation of shifts for points in the ring structure}

\item{sd2}{the standard deviation of shifts for points in the perturbed line segment structure}

\item{sd3}{the standard deviation of shifts for points in the 2-dimensional plane structure}

\item{d}{a number indicating the dimension of the generated data. d>3 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data matrix
\item clus - The vector of true cluster labels for each data point
}
}
\description{
Generating Manifold Mixture data
}
\examples{
X = MM_data()$data
}
