% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{NoiseYinyang_reg_data}
\alias{NoiseYinyang_reg_data}
\title{Generating Yinyang regression data}
\usage{
NoiseYinyang_reg_data(
  nnoise = 800,
  n_m = 400,
  n_c = 200,
  n_r = 2000,
  sd_c = 0.1,
  sd_r = 0.1,
  d = 2,
  sd_high = 0.1,
  sd_e = 0.1
)
}
\arguments{
\item{nnoise}{number of noisy points to add}

\item{n_m}{number of points in each of the two moon shapes. Positive integer value or vector with length=2}

\item{n_c}{number of points in each of the two small circles}

\item{n_r}{number of points in the outer ring}

\item{sd_c}{the standard deviation of shifts for points in the two small circles}

\item{sd_r}{the standard deviation of shifts for points in the outer ring}

\item{d}{a number indicating the dimension of the generated data. d>2 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}

\item{sd_e}{a number indicating the standard deviation for the additive Gaussian noises in the response}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data covariate matrix
\item Y - The generated responses for the covariates
}
}
\description{
Generating Yinyang regression data
}
\examples{
X = NoiseYinyang_reg_data()$data
plot(X[,1], X[,2])
}
