% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ring1000}
\alias{Ring1000}
\title{Simulated Ring Data with dimension 1000}
\format{
A list with 2 elements:
\describe{
\item{data}{A data frame with 1200 rows (observations) and 1000 columns (variables)}
\item{clus}{A vector of length 1200 specifying the true cluster memberships}
}
}
\usage{
Ring1000
}
\description{
Simulated Ring Data with dimension 1000
}
\keyword{datasets}
