% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{ring_data}
\alias{ring_data}
\title{Generating Ring data}
\usage{
ring_data(n_c = 200, n_r = 1000, sd_c = 0.1, sd_r = 0.1, d = 2, sd_high = 0.1)
}
\arguments{
\item{n_c}{a number indicating the number of points in the middle circle}

\item{n_r}{a number indicating the number of points in the outer ring}

\item{sd_c}{a number indicating the standard deviation of shifts for points on the middle circle}

\item{sd_r}{a number indicating the standard deviation of shifts for points on the outer ring}

\item{d}{a number indicating the dimension of the generated data. d>2 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data matrix
\item clus - The vector of true cluster labels for each data point
}
}
\description{
Generating Ring data
}
\examples{
X = ring_data()$data
plot(X[,1], X[,2])
}
