% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelBand}
\alias{skelBand}
\title{Calculating the base bandwidth for the S-kernel regressor}
\usage{
skelBand(centernn = NULL, px, skeleton)
}
\arguments{
\item{centernn}{the matrix where each row records the indices of the two closest knots of each knot.}

\item{px}{the projection proportion of the training set returned by the skelProject function.}

\item{skeleton}{a list returned by the voronSkeleton method constructing the skeleton graph.}
}
\value{
A numeric value for the global bandwidth to be used.
}
\description{
Calculating the base bandwidth for the S-kernel regressor
}
