% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelKernel}
\alias{skelKernel}
\title{Fitting the S-kernel regressor with fixed bandwidth}
\usage{
skelKernel(h1, testX, testSkeldists, trainY)
}
\arguments{
\item{h1}{the fixed bandwidth for kernel regression}

\item{testX}{the matrix of the test data points.}

\item{testSkeldists}{the matrix of skeleton-based distances between test and training points.}

\item{trainY}{the responses of the training points.}
}
\value{
A vector of predicted responses for the test points.
}
\description{
Fitting the S-kernel regressor with fixed bandwidth
}
