% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelKerneladopt}
\alias{skelKerneladopt}
\title{Fitting the S-kernel regressor with adaptive bandwidth}
\usage{
skelKerneladopt(hrate, testX, testSkeldists, trainY)
}
\arguments{
\item{hrate}{the scaling constant to apply to the adaptive calculated bandwidth}

\item{testX}{the matrix of the test data points.}

\item{testSkeldists}{the matrix of skeleton-based distances between test and training points.}

\item{trainY}{the responses of the training points.}
}
\value{
A vector of predicted responses for the test points.
}
\description{
Fitting the S-kernel regressor with adaptive bandwidth
}
