% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelknn}
\alias{skelknn}
\title{Fitting the S-kNN regressor}
\usage{
skelknn(k = 3, testX, testSkeldists, trainY)
}
\arguments{
\item{k}{the number of neighbors used in k-nearest-neighbor regression.}

\item{testX}{the matrix of the test data points.}

\item{testSkeldists}{the matrix of skeleton-based distances between test and training points.}

\item{trainY}{the responses of the training points.}
}
\value{
A vector of predicted responses for the test points.
}
\description{
Fitting the S-kNN regressor
}
