% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelProject}
\alias{skelProject}
\title{Project data points onto the skeleton}
\usage{
skelProject(nn = NULL, X, skeleton)
}
\arguments{
\item{nn}{the matrix where each row records the indices of the two closest centers of the corresponding data point.}

\item{X}{the matrix of the data points to be projected.}

\item{skeleton}{a list returned by the voronSkeleton method constructing the skeleton graph.}
}
\value{
A vector where each entry records the projection proportion of the corresponding data point on the edge by the two closest knots, from the smaller index knot. NA if the two closest knots not connected.
}
\description{
Project data points onto the skeleton
}
