% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelReg_runAll}
\alias{skelReg_runAll}
\title{Run all the skeleton-based methods with training and test datasets}
\usage{
skelReg_runAll(
  trainX,
  trainY,
  testX,
  seed = NULL,
  hrate_seq = c(1/8, 1/4, 1/2, 1, 2, 4, 6, 8, 10, 12, 14, 16),
  k_seq = seq(3, 36, by = 3),
  numknots = NULL,
  k_cut = 1,
  rep = 1000
)
}
\arguments{
\item{trainX}{a matrix/dataframe of the training covariates.}

\item{trainY}{a vector of the responses on training data.}

\item{testX}{a matrix/dataframe of the test covariates.}

\item{seed}{a random seed for skeleton construction.}

\item{hrate_seq}{a vector of scaling factors for bandwidth used for kernel regression.}

\item{k_seq}{a vector of the number of neighbors used for S-knn regressor.}

\item{numknots}{the number of knots to use for skeleton representation.}

\item{k_cut}{the number of disconnected components of the resulting skeleton graph.}

\item{rep}{number of repetitions for k-Means during skeleton construction process.}
}
\value{
A list with following components:
\describe{
\item{fits}{A list of vectors recording the predicted values of the data points by different methods.}
}
}
\description{
Run all the skeleton-based methods with training and test datasets
}
\examples{
\dontrun{
fits = skelReg_runAll(trainX,trainY,testX, seed = NULL,
                      hrate_seq = c(1,2),
                      k_seq = seq(3,6, by=3),
                      numknots=NULL,
                      k_cut=5,
                      rep = 1000)
}
}
