% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{skelReg_runAllCV}
\alias{skelReg_runAllCV}
\title{Run all the skeleton-based methods with cross-validation on a dataset}
\usage{
skelReg_runAllCV(
  X0,
  Y0,
  seed = NULL,
  nfold = 5,
  hrate_seq = c(1/8, 1/4, 1/2, 1, 2, 4, 6, 8, 10, 12, 14, 16),
  k_seq = seq(3, 36, by = 3),
  numknots = NULL,
  k_cut = 1,
  rep = 1000
)
}
\arguments{
\item{X0}{a matrix/dataframe of the covariates.}

\item{Y0}{a vector of the responses.}

\item{seed}{a random seed for skeleton construction.}

\item{nfold}{the number of folds for cross validation}

\item{hrate_seq}{a vector of scaling factors for bandwidth used for kernel regression.}

\item{k_seq}{a vector of the number of neighbors used for S-knn regressor.}

\item{numknots}{the number of knots to use for skeleton representation.}

\item{k_cut}{the number of disconnected components of the resulting skeleton graph.}

\item{rep}{number of repetitions for k-Means during skeleton construction process.}
}
\value{
A list with following components:
\describe{
\item{SEE}{A list recording the Sum of Squared Errors of the tested methods.}
\item{fits}{A list of vectors recording the predicted values of the data points by different methods.}
}
}
\description{
Run all the skeleton-based methods with cross-validation on a dataset
}
\examples{
\dontrun{
d = 100 #dimension of covariates
set.seed(1234)
traindata = Yinyang_reg_data(d = d) #simulate the Yinyang Regression data example
X0  = as.matrix(traindata$data)
Y0  = traindata$Y
results = skelReg_runAllCV(X0,Y0, seed = NULL, nfold = 5,
                           hrate_seq = c(1,2),
                           k_seq = c(3,6,9,12),
                           numknots=NULL,
                           k_cut=5,
                           rep = 1000)
}
}
