% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{SwissRollReg}
\alias{SwissRollReg}
\title{Generating SwissRoll Regression data}
\usage{
SwissRollReg(
  N = 2000,
  Height = 4 * pi,
  start = pi,
  roll = 1,
  d = 3,
  sd_high = 0.1,
  sd_e = 0.3
)
}
\arguments{
\item{N}{the sample size}

\item{Height}{the maximum radius}

\item{start}{the starting angle for the Swiss Roll}

\item{roll}{the number of rolls}

\item{d}{a number indicating the dimension of the generated data. d>3 dimensions are filled with independent Gaussian noises}

\item{sd_high}{a number indicating the standard deviation for Gaussian noises in the higher dimensions}

\item{sd_e}{a number indicating the standard deviation for the additive Gaussian noises in the response}
}
\value{
A list of generated data points and the true labels
\itemize{
\item data - The generated data covariate matrix
\item Y - The vector of generated response for each covariate
\item trueY - The vector of generated response without randomness
\item angles - The angle corresponding to each covariate
\item heights- The height of each corresponding covariate
}
}
\description{
Generating SwissRoll Regression data
}
\examples{
X = SwissRollReg()$data
}
