% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonCons.R
\name{vectorized_pdist}
\alias{vectorized_pdist}
\title{Vectorized pairwise Euclidean distance calculation}
\usage{
vectorized_pdist(A, B)
}
\arguments{
\item{A, B}{Matrices of data points with same number of columns}
}
\value{
A matrix of pairwise Euclidean distances
}
\description{
Vectorized pairwise Euclidean distance calculation
}
\examples{
A = matrix(c(1,2,3,4,5,6,7,8), ncol = 2, byrow = TRUE)
vectorized_pdist(A,A)
}
