% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonRegression.R
\name{voronSkeleton}
\alias{voronSkeleton}
\title{Construct the skeleton graph of a dataset based on Voronoi denisty weights}
\usage{
voronSkeleton(
  trainX,
  centers = NULL,
  labels = NULL,
  numknots,
  k_cut = 1,
  rep = 1000,
  seed = NULL
)
}
\arguments{
\item{trainX}{a matrix/dataframe of training data points.}

\item{centers}{an optional matrix/dataframe of the knots in the skeleton representation.}

\item{labels}{an optional numeric vector of skeleton membership labels for each data points.}

\item{numknots}{a number indicating the number of knots to use for skeleton representation.}

\item{k_cut}{a number indicating the number of resulting disconnected components in the skeleton representation.}

\item{rep}{a number indicating the repetition when overfitting kmeans to choose knots.}

\item{seed}{the random seed for the construction}
}
\value{
A list with following components:
\describe{
\item{centers}{The matrix of the skeleton knots.}
\item{labels}{The numeric vector indicating the cluser membership label of each center.}
\item{kdists}{The matrix of the pairwise Euclidean distances between the centers.}
\item{edgedists}{The matrix of edges in the skeleton with lengths. Entries in kdists not an edge set to 0.}
\item{nn}{The matrix where each row records the indices of the two closest centers of the corresponding data point.}
\item{g}{The igraph object for the constructed skeleton graph.}
}
}
\description{
Construct the skeleton graph of a dataset based on Voronoi denisty weights
}
